+++ src/MultiselectException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace AndreChalom\LaravelMultiselect;
+
+class MultiselectException extends \RuntimeException
+{
+}