+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 This file will keep a small log of changes to this package. This project adheres to Semantic Versioning.
 
-## [unrelased] - 2017-09-05
-Unrelased first draft of the package.
+## [0.0.2] - 2017-09-13
+
+- Provides a new exception to handle cases in which one of the  $default values is not included in the parameter $list. 
+- You can enable or disable a "strict" mode for generating the span element.
+- Several clarifications to the README.
+
+## [0.0.1] - 2017-09-06
+First draft of the package.